feat: Add componentLoaded flag for Blazor streaming
This commit adds a new variable `componentLoaded` to the BlazorStreamingUpdate class in the blazor-streaming.js file. This flag will be used to track whether the component has been loaded or not. Additionally, when the event "htmx:afterOnLoad" is triggered, it will call `htmx.process(document.body)` to process the document body.